requirements.txt
anyio==3.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-vision==3.3.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.1
python-telegram-bot==20.0
pytz==2022.7.1
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
urllib3==1.26.14
Werkzeug==2.2.2
